Selection function for population subset, modified random generator class so it can be called statically

+++ Tour.cpp
@@ -69,7 +69,7 @@ int Tour::getTourDistance() {
                 toCity = getCity(0);
             }
             // Get the distance between the two cities
-            tourDistance += originCity.getDistanceBetweenCities(toCity);
+            tourDistance += (int) originCity.getDistanceBetweenCities(toCity);
         }
         distance = tourDistance;
     }
